Refactor bundle deletion in IT

+++ test/integration_test.py
@@ -127,16 +127,17 @@ class IntegrationTest(AlwaysTearDownTestCase):
 
     def tearDown(self):
         self.set_lambda_test_mode(False)
-        self.delete_bundles()
-        # Delete again to test duplicate deletion notifications
-        self.delete_bundles(duplicates=True)
         super().tearDown()
 
     def test(self):
         if config.deployment_stage != 'prod':
-            self._test_indexing()
-            self._test_manifest()
-            self._test_drs()
+            try:
+                self._test_indexing()
+                self._test_manifest()
+                self._test_drs()
+            finally:
+                self._delete_bundles_twice()
+                self.assertTrue(self.project_removed_from_azul())
         self._test_other_endpoints()
 
     def _test_indexing(self):
@@ -200,17 +201,18 @@ class IntegrationTest(AlwaysTearDownTestCase):
         drs_endpoint = drs.drs_http_object_path(file_uuid)
         self.download_file_from_drs_response(self.check_endpoint_is_working(config.service_endpoint(), drs_endpoint))
 
-    def delete_bundles(self, duplicates=False):
-        if duplicates:
-            # Note: random.choices is with replacement (so the same choice may be made several times
-            notifications = random.choices(self.test_notifications, k=len(self.test_notifications) // 2)
-        else:
-            notifications = self.test_notifications
-        self.azul_client.delete_notification(notifications)
+    def _delete_bundles_twice(self):
+        self._delete_bundles(self.test_notifications)
+        # Delete again to test duplicate deletion notifications
+        # Note: random.choices is with replacement (so the same choice may be made several times
+        notifications = random.choices(self.test_notifications, k=len(self.test_notifications) // 2)
+        self._delete_bundles(notifications)
+
+    def _delete_bundles(self, notifications):
+        if notifications:
+            self.azul_client.delete_notification(notifications)
         self.wait_for_queue_level(empty=False)
         self.wait_for_queue_level(timeout=self.get_queue_empty_timeout(self.num_bundles))
-        self.assertTrue(self.project_removed_from_azul(),
-                        f"Project '{self.test_name}' was not fully removed from index within 5 min. of deletion")
 
     def set_lambda_test_mode(self, mode: bool):
         client = boto3.client('lambda')
